Move RemoteEventTimelineItem into its own module
